requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
openai>=1.0.0
python-dotenv>=0.19.0
streamlit>=1.0.0